Rename workerProfileStyles to workerInfoStyles
